Add NoItemsView and update list empty state
Introduced NoItemsView to display when the todo list is empty, including an animated 'Add Something' button. Updated ListView to show NoItemsView when there are no items and adjusted navigation bar items. Added SecondaryAccentColor asset for styling with transitions and animations to improve the overall UX.

+++ TodoList/Views/ListView.swift
@@ -11,27 +11,38 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
 
     var body: some View {
-        List{
-            ForEach(listViewModel.items) {item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack{
+            if listViewModel.items.isEmpty {
+               NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+                    
+            } else {
+                List{
+                    ForEach(listViewModel.items) {item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
+        //                Text(item.title)
+                        
                     }
-//                Text(item.title)
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(PlainListStyle())
                 
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
-        .listStyle(PlainListStyle())
+        
+        
         .navigationTitle("Todo List 📝")
         .navigationBarItems(
-            leading: EditButton(),
-            trailing: NavigationLink("Add", destination: {
-                AddView()
-            })
+            trailing: EditButton(),
+//            trailing: NavigationLink("Add", destination: {
+//                AddView()
+//            })
         )
     }
